refactor: move BuildTransaction method and related types to a separate file better organization
